Add HTTP/1.x request module
