+++ src/info/http1/mod.rs
@@ -5,3 +5,4 @@
 
 pub mod header;
 pub mod response;
+pub mod request;
